Fixed many bugs

+++ src/edu/kit/informatik/game/utility/Inputs.java
@@ -44,7 +44,7 @@ public class Inputs {
 
         input = askInput(shuffleSeedText, DIGIT_PATTERN);
         if (input.equals("quit")) return false;
-        shuffleSeed = Integer.parseInt(input);
+        shuffleSeed = Long.parseLong(input);
 
         return true;
     }
